- Removed unnecessary files. - Renamed a few files.
