+++ PyG/code/IsCyclic/iscyclic_weights.pt
